Move new BufferMethods to more commonly used file
